user edit fields are built (text only)

+++ db/migrate/20240622044219_create_client_answers.rb
@@ -3,7 +3,7 @@ class CreateClientAnswers < ActiveRecord::Migration[7.1]
     create_table :client_answers do |t|
       t.references :client_legal_form, null: false, foreign_key: true
       t.string :question
-      t.numeric :position
+      t.integer :position
       t.string :question_type
       t.string :options
       t.string :answer
